Extract hosted mutation journal lifecycles
