+++ crates/connect-hosted-provider/src/provider.rs
@@ -59,6 +59,9 @@ mod lifecycle;
 mod lifecycle_states;
 mod mutation_journal;
 pub use mutation_journal::HostedMutationJournalDiagnostics;
+mod mutation_journal_files;
+mod mutation_journal_migration;
+mod mutation_receipt;
 mod mutations;
 mod operation_context;
 mod operation_dispatch;
